src/requirements.txt
deepsearch-toolkit
numerize
pandas
tqdm
python-dateutil
countryinfo
openai